+++ src/Enums/PartyMandateScheme.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace FilamentAccounting\Enums;
+
+enum PartyMandateScheme: string
+{
+    case Core = 'CORE';
+    case B2b = 'B2B';
+}