+++ src/Enums/PartyMandateType.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace FilamentAccounting\Enums;
+
+enum PartyMandateType: string
+{
+    case OneOff = 'one_off';
+    case Recurring = 'recurring';
+}